Initial Hardcore PR + Custom Music Work

+++ src/map/zone.cpp
@@ -222,42 +222,42 @@ const int8* CZone::GetName()
     return (const int8*)m_zoneName.c_str();
 }
 
-uint8 CZone::GetSoloBattleMusic() const
+uint16 CZone::GetSoloBattleMusic() const
 {
     return m_zoneMusic.m_bSongS;
 }
 
-uint8 CZone::GetPartyBattleMusic() const
+uint16 CZone::GetPartyBattleMusic() const
 {
     return m_zoneMusic.m_bSongM;
 }
 
-uint8 CZone::GetBackgroundMusicDay() const
+uint16 CZone::GetBackgroundMusicDay() const
 {
     return m_zoneMusic.m_songDay;
 }
 
-uint8 CZone::GetBackgroundMusicNight() const
+uint16 CZone::GetBackgroundMusicNight() const
 {
     return m_zoneMusic.m_songNight;
 }
 
-void CZone::SetSoloBattleMusic(uint8 music)
+void CZone::SetSoloBattleMusic(uint16 music)
 {
     m_zoneMusic.m_bSongS = music;
 }
 
-void CZone::SetPartyBattleMusic(uint8 music)
+void CZone::SetPartyBattleMusic(uint16 music)
 {
     m_zoneMusic.m_bSongM = music;
 }
 
-void CZone::SetBackgroundMusicDay(uint8 music)
+void CZone::SetBackgroundMusicDay(uint16 music)
 {
     m_zoneMusic.m_songDay = music;
 }
 
-void CZone::SetBackgroundMusicNight(uint8 music)
+void CZone::SetBackgroundMusicNight(uint16 music)
 {
     m_zoneMusic.m_songNight = music;
 }
@@ -441,10 +441,10 @@ void CZone::LoadZoneSettings()
 
         inet_pton(AF_INET, (const char*)sql->GetData(1), &m_zoneIP);
         m_zonePort              = (uint16)sql->GetUIntData(2);
-        m_zoneMusic.m_songDay   = (uint8)sql->GetUIntData(3);           // background music (day)
-        m_zoneMusic.m_songNight = (uint8)sql->GetUIntData(4);           // background music (night)
-        m_zoneMusic.m_bSongS    = (uint8)sql->GetUIntData(5);           // solo battle music
-        m_zoneMusic.m_bSongM    = (uint8)sql->GetUIntData(6);           // party battle music
+        m_zoneMusic.m_songDay   = (uint16)sql->GetUIntData(3);          // background music (day)
+        m_zoneMusic.m_songNight = (uint16)sql->GetUIntData(4);          // background music (night)
+        m_zoneMusic.m_bSongS    = (uint16)sql->GetUIntData(5);          // solo battle music
+        m_zoneMusic.m_bSongM    = (uint16)sql->GetUIntData(6);          // party battle music
         m_tax                   = (uint16)(sql->GetFloatData(7) * 100); // tax for bazaar
         m_miscMask              = (uint16)sql->GetUIntData(8);
